Security fixes, bug fixes, enhancements
* Sanitized POST and GET inputs
* On get CSV file, added backup/restore if a new CSV file does not
download

+++ sheets2table/library/class-sheets2table-csv.php
@@ -36,7 +36,15 @@ Class S2T_CSV {
 	 * @var string $_file_path
 	 */
 	private $_file_path;
-		
+	/**
+	 * The part of the file name that indicates it is a backup
+	 *
+	 * @since 0.4.0
+	 * @access private
+	 * @var string $_backup_slug
+	 */
+	private $_backup_slug = "-backup.csv";
+	
 	/**
 	 * Instantiates the class with a CSV file
 	 *
@@ -47,8 +55,9 @@ Class S2T_CSV {
 	public function __construct($file_path) {
 	
 		# Nothing to do. Exit and let other errors fly
-		if ($file_path == "")
+		if ($file_path == "") {
 			return;
+		}
 		
 		# If the file doesn't exist, assume it is only the file name. 
 		# Prepend the dir to the file name
@@ -70,10 +79,11 @@ Class S2T_CSV {
 	 */
 	public function is_valid_file() {
 		
-		if (!file_exists($this->_file_path) || filesize($this->_file_path) <= 0)
+		if (!file_exists($this->_file_path) || filesize($this->_file_path) <= 0) {
 			return false;
-		else
-			return true;		
+		} else {
+			return true;
+		}
 	}
 	
 	/**
@@ -84,7 +94,67 @@ Class S2T_CSV {
 	 * @returns bool value of true if successfully deleted, otherwise returns false
 	 */
 	public function delete() {
-		return unlink($this->_file_path);
+		
+		if (file_exists($this->_file_path)) {
+			return unlink($this->_file_path);
+		}
+	}
+	
+	/**
+	 * Creates a copy of the CSV file as a backup
+	 *
+	 * @since 0.4.0
+	 *
+	 * @returns bool value of true if successfully backed up, otherwise returns false
+	 */
+	public function backup() {
+	
+		$result = false;
+	
+		if ($this->is_valid_file()) {
+	
+			# Append the new backup slug to the file and path
+			$backup_file_name = $this->_file_name . $this->_backup_slug;
+			$backup_file_path = $this->_file_path . $this->_backup_slug;
+			
+			$result = rename($this->_file_path, $backup_file_path);
+			
+			# If successful on the rename, set the updated file path
+			if ($result) {
+				$this->_file_path = $backup_file_path;
+				$this->_file_name = $backup_file_name;
+			}
+		}
+		
+		return $result;
+	}
+	
+	/**
+	 * Restores a backed up up CSV file
+	 *
+	 * @since 0.4.0
+	 *
+	 * @returns bool value of true if successfully restored, otherwise returns false
+	 */
+	public function restore() {
+		$result = false;
+	
+		if ($this->is_valid_file()) {
+	
+			# Remove the backup slug to the file and path
+			$restore_file_name = str_replace($this->_backup_slug, '', $this->_file_name);
+			$restore_file_path = str_replace($this->_backup_slug, '', $this->_file_path);
+			
+			$result = rename($this->_file_path, $restore_file_path);
+			
+			# If successful on the rename, set the updated file path
+			if ($result) {
+				$this->_file_path = $restore_file_path;
+				$this->_file_name = $restore_file_name;
+			}
+		}
+		
+		return $result;	
 	}
 	
 	/**
@@ -134,12 +204,13 @@ Class S2T_CSV {
 		# Display in human readable format
 		# 1048576 = 1024 * 1024 
 		 
-		if ($file_size >= 1048576 ) 
+		if ($file_size >= 1048576 ) {
 			$file_size = (int)($file_size / 1048576) . " MB"; 
-		elseif ($file_size >= 1024) 
+		} elseif ($file_size >= 1024) {
 			$file_size = (int)($file_size / 1024) . " KB"; 
-		else 
+		} else {
 			$file_size = $file_size . " bytes";
+		}
 
 		return $file_size;	
 	}
